修正x64编译失败
git-svn-id: http://qlanguage.googlecode.com/svn/trunk@198 6d982bb3-a5a7-393d-ffc5-582415be9c85

+++ Source/QLanguage/Parser/LRProduction.cpp
@@ -43,7 +43,7 @@ namespace QLanguage
     {
     }
 
-    LALR1Production::LALR1Production(const Production& p, uint pos) : parent(p), idx(pos)
+    LALR1Production::LALR1Production(const Production& p, size_t pos) : parent(p), idx(pos)
     {
     }
 
@@ -51,7 +51,7 @@ namespace QLanguage
     {
     }
 
-    LALR1Production::LALR1Production(const LALR1Production& p, uint pos) : parent(p), idx(pos), wildCards(p.wildCards)
+    LALR1Production::LALR1Production(const LALR1Production& p, size_t pos) : parent(p), idx(pos), wildCards(p.wildCards)
     {
     }
